Add intendation to statement blocks
All statements start on new lines, but nested code is not indented. Add
draft 1 of this through changing the E class to returning a reader with
the level of intendation. The Int in Reader is the number of spaces.

This is a simple solution to adding sugar to code-text orthogonally.